Move docker-run related functions into a docker_run.go file.
